Recent fixes toward release.

+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/statistic/ArrowheadFP.java
@@ -9,7 +9,7 @@ import edu.cmu.tetrad.graph.Graph;
  * The true positives are the number of arrowheads in both the true and estimated graphs.
  * Thus, if the true contains X*->Y and estimated graph either does not contain an edge from
  * X to Y or else does not contain an arrowhead at X for an edge from X to Y, one false
- * positive is counted. Similarly for false negatives.
+ * positive is counted. Similarly, for false negatives.
  *
  * @author jdramsey
  */
@@ -29,7 +29,7 @@ public class ArrowheadFP implements Statistic {
     @Override
     public double getValue(Graph trueGraph, Graph estGraph, DataModel dataModel) {
         ArrowConfusion confusion = new ArrowConfusion(trueGraph, estGraph);
-        return (double) confusion.getArrowsFp();
+        return confusion.getArrowsFp();
     }
 
     @Override
